Fix issue with sequence return types
Previously, the implementation called endpoint.GetSource() to get the application that the endpoint is in. This doesn't work for MOST endpoints, it seems to only be implemented for pubsub endpoints.
Refer to line 1837 at: https://github.com/anz-bank/sysl/blob/master/pkg/parse/listener_impl.go

Therefore it is necessary to pass in the appname into the CreateReturnDataModel function